requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
bleach==6.0.0
Django==4.1.6
django-filter==22.1
django-guardian==2.4.0
django-phonenumber-field==6.3.0
django-rest-passwordreset==1.2.1
djangorestframework==3.14.0
drf-api-logger==1.1.12
importlib-metadata==6.0.0
Markdown==3.4.1
phonenumbers==8.12.53
Pillow==9.2.0
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
webencodings==0.5.1
zipp==3.13.0
